fix the return type of the intermediate builder methods in datasources commands
The intermediate builder methods (i.e., those that return the builder instance
so that other methods can be called) used to always return the type Builder,
even if the actual builder instance was a subclass. The correct solution
would be to return a "self type", but sadly, Java doesn't support self types.

An easy solution is to abuse generics: the parent Builder class has a type
parameter THIS that is always of the "current" Builder class. Then, all
the intermediate methods return THIS. The type parameter plays the role
of the self type, without a lot of overhead.