replace the OnlineOptions.serverType method with .protocol
The serverType name is slightly misleading -- the only thing it affects
is the communication protocol (remoting or http-remote). When localDefault()
is used, it also affects the port number, but that is kinda implied by
the protocol. So a method called protocol (and a new enum called
ManagementProtocol) is more appropriate.

The serverType method and the ServerType enum are now deprecated.