support "reload" in OnlineMamagementClient.executeCli
The CLI requires a special implementation of ModelControllerClient
for "reload" and "shutdown", but Creaper uses the standard one, so
these operations don't work.

Hhowever, "reload" with no options is particularly common in CLI scripts,
so this commit adds a special case to help in those situations. Note
that "reload" with options (e.g. "reload --admin-only=true") or "shutdown"
(with or without options) are not supported.