fix Administration.reload when the server is in "restart-required"
Calling Administration.reload when the server is in state "restart-required"
used to timeout instead of returning normally. This was caused by the method
waiting for the server to get into the state "running". However, reloading
a server that is in "restart-required" leads directly to "restart-required".
This commit fixes the problem by accounting for that possibility.