add a workaround for WFCORE-1082
The CLI operation "module add" generates the module.xml file automatically
if it isn't provided. However, it generates the <main-class> element wrongly,
using the attribute "value" instead of "name". This is fixed since WildFly
Core 2.0.0, but the AddModule command is useful way before that. So this
commit adds a workaround to the AddModule class which does a simple textual
replace of the wrong string.