avoid exceptions when Operations.removeIfExists is called from command
When calling Operations.removeIfExists outside of an OnlineCommand, it returns
false correctly when the desired management resource doesn't exist. However,
when it's called from an OnlineCommand, it throws an exception. The reason
is that inside a command, the management client is wrapped in a proxy that
detects failed management operations and throws exceptions automatically,
so that commands don't have to do that.

This is an important convenience that needs to be retained. However, sometimes
it's simply expected that an operation will fail (Operations.removeIfExists is
one case, Operations.exists is another and one could argue that this can happen
for user-constructed operations as well). So there needs to be a way
to communicate such intent from inside the command. If we were on Java 8,
a lambda would be natural, but on Java 6, this is simply too unwieldy. Instead,
a new method allowFailures is added to OnlineManagementClient that returns
a Closeable. Calling that method and closing its result delimit the block of
code in which failures shouldn't be converted to exceptions because they are
expected. This is still somewhat wordy in Java 6, but in Java 7 and above,
try-with-resources can be used to bring the syntactic overhead down.