add OfflineManagementClient.serverVersion
This commit (finally!) implements server version discovery in offline case.
The version is exposed as OfflineManagementClient.serverVersion, just like
in OnlineManagementClient.

Technically, it's a breaking change (see the changes in the tests), because
previsouly, users were able to use OfflineManagementClient to transform
any and all XML files. This is now disallowed, because the XML file must
contain an AS7/WildFly management version. That, however, has always
been the intent, so the number of affected users should be low.