add ServerVersion and deprecate ManagementVersion
The ManagementVersion enum serves well as a repository of known management
versions, but it's really terrible when used to represent current version
of the server, which is potentially unknown. This means that for each new
server version, the enum has to be modified and a new release made. It is now
clear that, in the hindsight, this was a bad design decision and must be
changed.

This commit introduces the class ServerVersion which is not an enum
but a regular class that can be instantiated at will. It still contains
all known versions of the server as a bunch of constants, but it also
has a factory method that can be used to instantiate an object to represent
current server version (which, again, can be unknown).

The ManagementVersion enum is now deprecated, ServerVersion being
the suggested replacement.

This commit doesn't fix the deprecations yet, this is expected
in a follow-up commit.