add OfflineManagementClient.version()
This commit (finally!) implements server version discovery in offline case.
The version is exposed as OfflineManagementClient.version() and also as
OfflineCommandContext.version, just like in the online case.

Technically, it's a breaking change (see the changes in the tests), because
previsouly, users were able to use OfflineManagementClient to transform
any and all XML files. This is now disallowed, because the XML file must
be an AS7/WildFly configuration file and must contain a management version.
That, however, has always been the intent, so the number of affected users
should be low.

An interesting note is that for certain AS7/EAP6 versions, the code needs
to normalize schema version numbers from XML files to management version
numbers, so that offline and online versions are consistent. This will likely
never matter, because the inconsistency only exists in very old releases.