fix the treatment of default host in managed domain
With the offline client, the notion of default host doesn't make
any sense, because the client always works with a single file.

With online client, default host is used to transform operations
on addresses /core-service=... to /host=.../core-service=..., and
if there is no default host to prepend, an exception is thrown.
That, however, is wrong, because there are legitimate uses
of operations on /core-service=... in managed domain, e.g. RBAC.
The solution is simple -- if there's no default host defined,
the operation will be executed unchanged.