update GroovyXmlTransform to understand new first-level XML elements
WildFly added new elements to the set of possible first-level elements
in the XML configuration files. This commit updates the Subtree class
to understand them (which requires making a difference between host.xml
and domain.xml, as host.xml is suddenly very similar to standalone.xml).

Also, to maintain correct order of the XML elements, it is no longer
sufficient to maintain a single list which blends all the possibilities,
so the list was split into three separate lists -- for the "domain" root
element, for "host" and for "server".