make management client libraries <scope>provided</scope>
A long time ago, it might have been useful for Creaper to automatically
propagate all management client libraries (*-controller-client, *-cli
and also wildlfy-patching). That was possible only because there was
a single version of the application server Creaper had to handle.
Now that we have AS 7 / EAP 6, multiple WildFly versions and EAP 7 upcoming,
it's become more of a hassle. Users should provide a correct version of all
the client libraries themselves. Hence, <scope>provided</scope>.