remove debugging code from the Administration class
The Administration class used to contain a piece of debugging code.
This is no longer necessary and this commit removes it. Sadly,
some methods in the class didn't declare checked exceptions properly.
Adding them would be source-incompatible, which would require
new major release. That is currently undesirable. Instead, this commit
preserves the most awful part of the debugging code (the "sneaky throw"
mechanism) so that no incompatible change has to be made.