add a CliScript command
In addition to existing CliFile command that reads a CLI script
from a file or classpath resource, this commit adds a CliScript
command that reads the CLI script from a string. The CliFile
command has been refactored to call CliScript under the hood.

Also, the CliFile has one new constructor that takes a class
as a resource loader and a separate classpath resource path
(similarly to GroovyXmlTransform).