Replace wildfly-logging with jboss-logging
In this commit [1] we (apart from other changes) moved from the
`org.jboss.logging:jboss-logging` artifact to the
`org.wildfly.core:wildfly-logging` artifact. After some time it looks
that this change wasn't correct. These two aren't exactly replacements
to each other. On top of that we ran into issue where on JDK17 the
creaper isn't able to connect to the WildFly 27+. This could be resolved
by excluding all transitive dependencies brought in by `wildfly-logging`
but more correct way is probably to move back to the `jboss-logging`
artifact now.

[1] ba6d8c5c8a741d99254a722a70c3808af1da83c9