[#244] drop explicit dependency on jboss-logging
fixes https://github.com/wildfly-extras/creaper/issues/244

It attempts to fix issue due to current dependecy situation:
* creaper has explicit dependency on jboss-logging 3.3.0.Final
* jboss-logging added new methods in 3.6.0
* wildfly-core uses new methods from jboss-logging 3.6.x since 26.0.0.Beta3 which means WildFly 34
* it results in #244

Let's bring in jboss-logging transitivitely from wildfly-core to keep version aligned.

----

Originally jboss-logging dependency was added with this commit message

```
OnlineManagementClient and OfflineManagementClient now log all commands
and operations they are performing. Operations (online only) are logged
on the DEBUG level, commands are logged on INFO. Also, commands are now
advised to provide a toString() method that is used for logging.

The logging framework that is used in Creaper is JBoss Logging, because:

1. It is not actually a logging framework, it's only a facade that delegates
   the logging work to some logging framework that appears to be used.
2. It can detect all commonly used logging frameworks (Logback, Log4j, JUL)
   and one uncommon too (JBoss LogManager).
3. Creaper already depended on it transitively. Avoiding new dependencies
   is good.
4. It has a really small and nice API.
```